Move the creatEmptyOrder of file.
