+++ src/context/OrderContext.spec.tsx
@@ -1,10 +1,11 @@
 import {act, fireEvent, render, screen, waitFor} from '@testing-library/react';
 import {jest} from '@jest/globals';
-import {createEmptyOrder, Order} from '../api/orders/Orders';
+import {Order} from '../api/orders/Orders';
 import {useOrderContext, OrderProvider} from './OrderContext';
 import type {OrderContextType} from '../@types/order';
 import {storage} from 'src/utils/localStorage';
 import {OrderStatus, PaymentMethod} from 'src/@types/order';
+import createEmptyOrder from 'src/api/orders/createEmptyOrder';
 
 function DummyComponent({futureOrder}: {futureOrder?: Order}) {
   const value = useOrderContext();
